# backend/internal/httpapi/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do LogN. Erro de rota que o app lê sai como {\"code\", \"message\"}, e o código é contrato.",
        "title": "LogN API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/waitlist": {
            "post": {
                "description": "Formulário da landing: `Origin` tem de ser a landing, ou `Sec-Fetch-Site` same-site. Todo desfecho, inclusive erro e limite, é 303 para uma página da landing.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "Entra na lista de espera",
                "parameters": [
                    {
                        "type": "string",
                        "description": "E-mail",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "pt-BR",
                            "en",
                            "es"
                        ],
                        "type": "string",
                        "description": "Língua",
                        "name": "locale",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Honeypot; tem de vir vazio",
                        "name": "website",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "303": {
                        "description": "redireciona para thanks ou error"
                    }
                }
            }
        },
        "/api/v1/waitlist/confirm": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "Página do botão de um link do e-mail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token assinado do link",
                        "name": "t",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "página com o botão",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "303": {
                        "description": "link inválido ou já confirmado"
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "waitlist"
                ],
                "summary": "Confirma ou sai da lista de espera",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token assinado do link",
                        "name": "t",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "303": {
                        "description": "redireciona para confirmed, left ou error"
                    }
                }
            }
        },
        "/api/v1/waitlist/leave": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "waitlist"
                ],
                "summary": "Página do botão de um link do e-mail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token assinado do link",
                        "name": "t",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "página com o botão",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "303": {
                        "description": "link inválido ou já confirmado"
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "waitlist"
                ],
                "summary": "Confirma ou sai da lista de espera",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token assinado do link",
                        "name": "t",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "303": {
                        "description": "redireciona para confirmed, left ou error"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token do login, no formato `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}